requirements.txt
attrs==23.2.0
branca==0.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
dnspython==2.6.1
fiona==1.9.6
folium==0.16.0
geopandas==0.14.3
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pyarrow==15.0.2
pymongo==4.6.3
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
shapely==2.0.4
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
xyzservices==2024.4.0
zipp==3.18.1
